requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-telegram-bot==21.0.1
requests==2.28.1
openai==0.27.8
python-dotenv==0.21.0
python-magic==0.4.27
beautifulsoup4==4.12.2
nltk==3.8.1
fasttext-wheel==0.9.2
flashtext==2.7
gensim==4.3.1
hazm==0.10.0
